feat: Firmware: Read temperatures

Close #26

+++ firmware/src/devices/click.cpp
@@ -71,7 +71,7 @@ void ClickSpan::copy(ClickSpan* timespan)
 
 
 
-Click::Click (int num)
+Click::Click (uint8_t num)
 {
     this->m_num = num;
     this->m_counter = 0;
