+++ firmware/src/include/click.h
@@ -37,7 +37,7 @@ class ClickSpan
 class Click
 {
     public:
-        Click (int num);
+        Click (uint8_t num);
         virtual ~Click();
 
         enum Type {
@@ -54,7 +54,7 @@ class Click
         Type type();
 
     private:
-        int m_num;
+        uint8_t m_num;
         unsigned long m_counter;
         Type m_type;
